src/terminator/terminal/PtyProcess.java: src/terminator/view/JTerminalPane.java: we're not aware of any system with 64-bit pid_t (quite the opposite: the 64-bit systems we know of have a 32-bit pid_t, and 64-bit Linux still has a default maximum pid of 32768), but if we're publishing this interface (as we intend to) we probably want the widest signed integral type available.

+++ terminator/trunk/src/terminator/terminal/PtyProcess.java
@@ -44,7 +44,7 @@ public class PtyProcess {
     }
     
     private int fd = -1;
-    private int pid;
+    private long pid;
     private String slavePtyName;
     
     private boolean didDumpCore = false;
@@ -123,7 +123,7 @@ public class PtyProcess {
         return outStream;
     }
     
-    public int getPid() {
+    public long getPid() {
         return pid;
     }
